kafka-kinesis-bridge/requirements.txt
kafka-python-ng>=2.2.2
boto3>=1.34.0